add tab design
