+++ app/src/main/res/layout/fragment_movie_lists.xml
@@ -12,10 +12,15 @@
         android:id="@+id/mlf_tab_layout"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
+        android:background="@color/background"
         app:layout_constraintBottom_toTopOf="@id/mlf_pager"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        app:tabSelectedTextColor="@color/white"
+        app:tabTextColor="@color/mlf_tab_text"
+        app:tabIndicatorColor="@color/genres"
+        app:tabTextAppearance="@style/TextAppearance.Design.Tab.MoviesTab"/>
 
 
     <androidx.viewpager2.widget.ViewPager2
