+++ app/src/main/res/values/style.xml
@@ -9,4 +9,11 @@
         <item name="android:textStyle">bold</item>
         <item name="android:textColor">@color/cast_name</item>
     </style>
+
+    <style name="TextAppearance.Design.Tab.MoviesTab" parent="TextAppearance.Design.Tab">
+        <item name="textAllCaps">false</item>
+        <item name="android:textAllCaps">false</item>
+        <item name="android:textSize">14sp</item>
+    </style>
+
 </resources>
